x_1/momoweb/requirements.txt
Django==3.1.3
beautifulsoup4==4.9.3
fake-useragent==0.1.11
lxml==4.6.1
requests==2.25.0
python-dateutil==2.8.1
djangorestframework==3.12.2
django-crontab==0.7.1
demjson==2.2.4
